% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nclRmodelling-package.R
\docType{package}
\name{jrModelling-package}
\alias{jrModelling-package}
\alias{jrModelling}
\title{Statistical Modelling by Jumping Rivers}
\description{
Functions and datasets used in external R courses
}
\keyword{package}
