% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{MvNormal}
\alias{MvNormal}
\title{Multivariate normal}
\usage{
MvNormal(n = 1000, cor = 0.8)
}
\arguments{
\item{n}{number of point in scatter}

\item{cor}{correlation between the two variables}
}
\description{
Function for generating a plot of bivariate normal with a given correlation 
between 2 variables. Created for the purpose of generating correlation plots in the
handout material. Not intended for other purposes.
}
