% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{my_plot}
\alias{my_plot}
\title{Custom plot}
\usage{
my_plot(xy, xlim = c(-4, 4), ylim = c(-4, 4), eps = 1e-15)
}
\arguments{
\item{xy}{matrix of xy values}

\item{xlim}{numeric vector for x axis range}

\item{ylim}{numeric vector for y axis range}

\item{eps}{numeric tolerance for which cor not defined (don't divide by zero)}
}
\description{
A customised plot used for generating the correlation
graphics in the handout material.
Not intended for other purposes.
}
