% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{others}
\alias{others}
\title{Other correlation plots}
\usage{
others(n = 1000)
}
\arguments{
\item{n}{number of points in scatter}
}
\description{
Generate scatter plots for handout material and slides. Gives plots
of correlations with assorted shapes. Created for the purpose of handout
material. Not intended for any other purpose
}
