% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_example.R
\name{regression_example}
\alias{regression_example}
\title{Regression with high leverage example}
\usage{
regression_example(includecode = FALSE)
}
\arguments{
\item{includecode}{Logical. If TRUE the code
included in the ui.R and server.R source files will
be shown alongside the app.}
}
\description{
This function runs an example of a shiny application to
do with a regression example to examine the change in
linear model fit when removing points of high influence.
}
