% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{rot_normal}
\alias{rot_normal}
\title{Rotated normal scatter}
\usage{
rot_normal(n = 1000, t = pi/2)
}
\arguments{
\item{n}{number of points in scatter}

\item{t}{angular rotation to apply}
}
\description{
For a bivariate normal apply a rotation and plot resultant scatter.
Created for the purpose of correlation plots in handout material.
Not intended for any other purpose
}
