% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{zodiac}
\alias{zodiac}
\title{Star signs of CEOs}
\value{
A data frame
}
\description{
Star signs of over 4000 business CEOS
were taken
}
\examples{
data(zodiac, package="jrModelling")
}
\keyword{datasets}
